Remove all translations
While the app name for Aurora Store's Priviliged Extension was changed, it was only done so for English,
however, this means that other languages mostly all reference this app as F-Droid's Priviliged Extension.
This commits removes all other languages.

It may be necessary to do extra work to make this work I'm unaware of, in that case it makes sense to reject this,
since the Priviliged Extensions overall are deprecated.

Change-Id: I4b75190741cd971b2e1b8e56cb65e13580d1428e